+++ core-navigation/lib-core-navigation/src/main/java/ru/surfstudio/android/core/ui/navigation/activity/navigator/ActivityNavigator.java
@@ -42,9 +42,9 @@ import ru.surfstudio.android.core.ui.event.lifecycle.completely.destroy.OnComple
 import ru.surfstudio.android.core.ui.event.lifecycle.pause.OnPauseDelegate;
 import ru.surfstudio.android.core.ui.event.lifecycle.resume.OnResumeDelegate;
 import ru.surfstudio.android.core.ui.event.newintent.NewIntentDelegate;
-import ru.surfstudio.android.core.ui.event.result.BaseActivityResultDelegate;
-import ru.surfstudio.android.core.ui.event.result.CrossFeatureSupportOnActivityResultRoute;
-import ru.surfstudio.android.core.ui.event.result.SupportOnActivityResultRoute;
+import ru.surfstudio.android.core.ui.navigation.event.result.BaseActivityResultDelegate;
+import ru.surfstudio.android.core.ui.navigation.event.result.CrossFeatureSupportOnActivityResultRoute;
+import ru.surfstudio.android.core.ui.navigation.event.result.SupportOnActivityResultRoute;
 import ru.surfstudio.android.core.ui.navigation.ActivityRouteInterface;
 import ru.surfstudio.android.core.ui.navigation.Navigator;
 import ru.surfstudio.android.core.ui.navigation.ScreenResult;