+++ core-navigation/lib-core-navigation/src/main/java/ru/surfstudio/android/core/ui/navigation/feature/route/dynamic_feature/ActivityDynamicCrossFeatureWithResultRoute.kt
@@ -18,8 +18,8 @@
 package ru.surfstudio.android.core.ui.navigation.feature.route.dynamic_feature
 
 import android.content.Intent
-import ru.surfstudio.android.core.ui.event.result.CrossFeatureSupportOnActivityResultRoute
-import ru.surfstudio.android.core.ui.event.result.SupportOnActivityResultRoute
+import ru.surfstudio.android.core.ui.navigation.event.result.CrossFeatureSupportOnActivityResultRoute
+import ru.surfstudio.android.core.ui.navigation.event.result.SupportOnActivityResultRoute
 import java.io.Serializable
 
 /**