+++ permission/lib-permission/src/main/java/ru/surfstudio/android/core/ui/permission/screens/settings_rational/DefaultSettingsRationalActivity.kt
@@ -6,7 +6,7 @@ import android.os.Bundle
 import android.provider.Settings
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
-import ru.surfstudio.android.core.ui.R
+import ru.surfstudio.android.core.ui.permission.R
 import ru.surfstudio.android.core.ui.navigation.Route
 
 private const val SETTINGS_REQUEST_CODE = 1005