+++ core-navigation/RELEASE_NOTES.md
@@ -3,3 +3,5 @@
 ## 0.5.0-alpha
 ##### Core-navigation
 * ANDDEP-1049 All classes responsible for navigation migrated from `core-ui` to `core-navigation`
+* **NO BACKWARD COMPATIBILITY** ANDDEP-1049 Package for `BaseActivityResultDelegate`,`SupportOnActivityResultRoute` and `CrossFeatureSupportOnActivityResultRoute` is changed. 
+from `ru.surfstudio.android.core.ui.event.result` to:  `ru.surfstudio.android.core.ui.navigation.event.result`